Make migrations and login page in frontend

+++ database/migrations/2022_05_16_1652697490_create_mh_esubmit_file_designation_table.php
@@ -11,7 +11,7 @@ class CreateMhEsubmitFileDesignationTable extends Migration
         Schema::create('mh_esubmit_file_designation', function (Blueprint $table) {
 
         $table->id();
-		$table->integer('journalID');
+		$table->unsignedBigInteger('journalID');
 		$table->string('title');
 		$table->datetime('entryDate');
 		$table->tinyInteger('status');
