+++ database/migrations/2022_05_16_1652699019_create_mh_journals_checklist_text_table.php
@@ -11,7 +11,7 @@ class CreateMhJournalsChecklistTextTable extends Migration
         Schema::create('mh_journals_checklist_text', function (Blueprint $table) {
 
             $table->id();
-            $table->integer('journalID');
+            $table->unsignedBigInteger('journalID');
             $table->text('description');
             $table->timestamps();
         });
