+++ database/migrations/2022_05_16_1652700474_create_mh_sms_aedecisions_table.php
@@ -12,8 +12,8 @@ class CreateMhSmsAedecisionsTable extends Migration
 
 			$table->id();
 			$table->string('editorType');
-			$table->integer('editorID');
-			$table->integer('userID');
+			$table->unsignedBigInteger('editorID');
+			$table->unsignedBigInteger('userID');
 			$table->string('attachment');
 			$table->string('emailFrom');
 			$table->string('emailTo');
