+++ database/migrations/2022_05_16_1652700559_create_mh_sms_aedecisions_associate_table.php
@@ -12,8 +12,8 @@ class CreateMhSmsAedecisionsAssociateTable extends Migration
 
 			$table->id();
 			$table->string('editorType');
-			$table->integer('editorID');
-			$table->integer('userID');
+			$table->unsignedBigInteger('editorID');
+			$table->unsignedBigInteger('userID');
 			$table->string('attachment');
 			$table->string('emailFrom');
 			$table->string('emailTo');
