+++ database/migrations/2022_05_17_1652761701_create_mh_sms_orignalityreports_table.php
@@ -12,10 +12,9 @@ class CreateMhSmsOrignalityreportsTable extends Migration
 
             $table->id();
 
-		$table->integer('editorialID');
-		$table->integer('editorID');
-		$table->integer('journalID');
-		;
+		$table->unsignedBigInteger('editorialID');
+		$table->unsignedBigInteger('editorID');
+		$table->unsignedBigInteger('journalID');
 		$table->text('comments');
 		$table->string('orderNumber');
 		$table->datetime('entryDate');
