+++ database/migrations/2022_05_17_1652761778_create_mh_sms_pdfcontent_table.php
@@ -12,7 +12,7 @@ class CreateMhSmsPdfcontentTable extends Migration
 
             $table->id();
 
-            $table->integer('userID');;;
+            $table->unsignedBigInteger('userID');;;
             $table->string('meta_content_file');
             $table->string('meta_table_file');
             $table->string('orderNumber');
