+++ database/migrations/2022_05_17_1652762134_create_mh_sms_selected_ass_editors_table.php
@@ -12,7 +12,7 @@ class CreateMhSmsSelectedAssEditorsTable extends Migration
 
             $table->id();
 
-		$table->integer('assEditorID');
+		$table->unsignedBigInteger('assEditorID');
 		$table->string('firstName');
 		$table->string('emailAddress');
 		$table->string('orderNumber');
