+++ database/migrations/2022_05_17_1652762314_create_mh_sms_typetitleabstract_table.php
@@ -11,7 +11,7 @@ class CreateMhSmsTypetitleabstractTable extends Migration
         Schema::create('mh_sms_typetitleabstract', function (Blueprint $table) {
 
             $table->id();
-            $table->integer('userID');
+            $table->unsignedBigInteger('userID');
             $table->integer('manuscriptID');
             $table->integer('typeOfManuScript');
             $table->integer('areaOfSpecificInterest');
