+++ database/migrations/2022_05_17_1652762364_create_mh_sms_unsuggestedrev_table.php
@@ -12,7 +12,7 @@ class CreateMhSmsUnsuggestedrevTable extends Migration
 
             $table->id();
 
-		$table->integer('userID');
+		$table->unsignedBigInteger('userID');
 		$table->integer('manuscriptID');
 		$table->string('recommendingName');
 		$table->string('recommendingEmail');
